[FLINK-15954] Reuse RawSerializer
This commit extracts the RawSerializer and reuses it
across the multiplexed state accessors.
The TaggedSerializer that is removed as part of this commit
is identical to the RawSerializer when initialized with an empty
tag (which we did)
