+++ statefun-flink/statefun-flink-core/src/main/java/org/apache/flink/statefun/flink/core/state/MultiplexedMapStateAccessor.java
@@ -17,12 +17,9 @@
  */
 package org.apache.flink.statefun.flink.core.state;
 
-import java.io.IOException;
 import java.util.Objects;
 import org.apache.flink.api.common.state.MapState;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
-import org.apache.flink.core.memory.DataInputDeserializer;
-import org.apache.flink.core.memory.DataOutputSerializer;
 import org.apache.flink.statefun.flink.core.generated.MultiplexedStateKey;
 import org.apache.flink.statefun.sdk.state.Accessor;
 
@@ -73,31 +70,6 @@ final class MultiplexedMapStateAccessor<T> implements Accessor<T> {
       mapStateHandle.remove(accessorMapKey);
     } catch (Exception e) {
       throw new RuntimeException(e);
-    }
-  }
-
-  private static final class RawSerializer<T> {
-    private final TypeSerializer<T> delegate;
-    private final DataOutputSerializer output;
-    private final DataInputDeserializer input;
-
-    RawSerializer(TypeSerializer<T> delegate) {
-      this.delegate = Objects.requireNonNull(delegate);
-      this.output = new DataOutputSerializer(32);
-      this.input = new DataInputDeserializer();
-    }
-
-    byte[] serialize(T value) throws IOException {
-      output.clear();
-      delegate.serialize(value, output);
-      return output.getCopyOfBuffer(); // TODO: consider avoiding buffer copying
-    }
-
-    T deserialize(byte[] bytes) throws IOException {
-      input.setBuffer(bytes);
-      final T value = delegate.deserialize(input);
-      input.releaseArrays();
-      return value;
     }
   }
 }
